Move case type and datestamp
