+++ app/presenters/summary/sections/case_details.rb
@@ -10,17 +10,17 @@ module Summary
         [
           Components::ValueAnswer.new(
             :case_type, kase.case_type,
-            change_path: edit_steps_case_case_type_path
+            change_path: edit_steps_client_case_type_path
           ),
 
           Components::DateAnswer.new(
             :appeal_lodged_date, kase.appeal_lodged_date,
-            change_path: edit_steps_case_appeal_details_path
+            change_path: edit_steps_client_appeal_details_path
           ),
 
           Components::FreeTextAnswer.new(
             :appeal_with_changes_details, kase.appeal_with_changes_details,
-            change_path: edit_steps_case_appeal_details_path
+            change_path: edit_steps_client_appeal_details_path
           ),
 
           # This is an optional field, depending on case type
@@ -28,7 +28,7 @@ module Summary
           Components::FreeTextAnswer.new(
             :previous_maat_id, kase.appeal_maat_id,
             show: !kase.appeal_maat_id.nil?,
-            change_path: edit_steps_case_appeal_details_path
+            change_path: edit_steps_client_appeal_details_path
           ),
 
           Components::FreeTextAnswer.new(
