+++ app/services/decisions/case_decision_tree.rb
@@ -1,15 +1,8 @@
 module Decisions
-  # rubocop:disable Metrics/ClassLength
   class CaseDecisionTree < BaseDecisionTree
     # rubocop:disable Metrics/MethodLength, Metrics/CyclomaticComplexity, Metrics/AbcSize
     def destination
       case step_name
-      when :case_type
-        after_case_type
-      when :appeal_details
-        date_stamp_if_needed
-      when :date_stamp
-        edit(:urn)
       when :urn
         charges_summary_or_edit_new_charge
       when :charges
@@ -41,20 +34,6 @@ module Decisions
     # rubocop:enable Metrics/MethodLength, Metrics/CyclomaticComplexity, Metrics/AbcSize
 
     private
-
-    def after_case_type
-      return edit(:appeal_details) if form_object.case_type.appeal?
-
-      date_stamp_if_needed
-    end
-
-    def date_stamp_if_needed
-      if DateStamper.new(form_object.crime_application, form_object.case.case_type).call
-        edit(:date_stamp)
-      else
-        edit(:urn)
-      end
-    end
 
     def charges_summary_or_edit_new_charge
       return edit(:charges_summary) if case_charges.any?
@@ -138,5 +117,4 @@ module Decisions
       current_charge.offence_dates.map(&:date_from).exclude?(nil)
     end
   end
-  # rubocop:enable Metrics/ClassLength
 end
