+++ app/services/decisions/client_decision_tree.rb
@@ -6,7 +6,15 @@ module Decisions
       when :has_partner
         after_has_partner
       when :details
-        after_client_details
+        edit(:case_type)
+      when :case_type
+        after_case_type
+      when :appeal_details
+        date_stamp_if_needed
+      when :date_stamp
+        start_address_journey(HomeAddress)
+      when :contact_details
+        after_contact_details
       when :has_nino
         edit(:benefit_type)
       when :benefit_type
@@ -17,8 +25,6 @@ module Decisions
         after_dwp_check
       when :has_benefit_evidence
         after_has_benefit_evidence
-      when :contact_details
-        after_contact_details
       else
         raise InvalidStep, "Invalid step '#{step_name}'"
       end
@@ -36,12 +42,34 @@ module Decisions
       end
     end
 
-    def after_client_details
-      if current_crime_application.age_passported?
+    def after_case_type
+      return edit(:appeal_details) if form_object.case_type.appeal?
+
+      date_stamp_if_needed
+    end
+
+    def date_stamp_if_needed
+      if DateStamper.new(form_object.crime_application, form_object.case.case_type).call
+        edit(:date_stamp)
+      else
         start_address_journey(HomeAddress)
+      end
+    end
+
+    def after_contact_details
+      if form_object.correspondence_address_type.other_address?
+        start_address_journey(CorrespondenceAddress)
+      elsif current_crime_application.age_passported?
+        edit('/steps/case/urn')
       else
         edit(:has_nino)
       end
+    end
+
+    def start_address_journey(address_class)
+      address = address_class.find_or_create_by(person: applicant)
+
+      edit('/steps/address/lookup', address_id: address)
     end
 
     def after_benefit_type
@@ -67,29 +95,15 @@ module Decisions
     end
 
     def after_dwp_check
-      start_address_journey(HomeAddress)
+      edit('/steps/case/urn')
     end
 
     def after_has_benefit_evidence
       if form_object.has_benefit_evidence.yes?
-        start_address_journey(HomeAddress)
+        edit('/steps/case/urn')
       else
         show(:evidence_exit)
       end
-    end
-
-    def after_contact_details
-      if form_object.correspondence_address_type.other_address?
-        start_address_journey(CorrespondenceAddress)
-      else
-        edit('/steps/case/case_type')
-      end
-    end
-
-    def start_address_journey(address_class)
-      address = address_class.find_or_create_by(person: applicant)
-
-      edit('/steps/address/lookup', address_id: address)
     end
 
     def applicant
