+++ app/views/steps/client/date_stamp/edit.html.erb
