+++ config/locales/en/helpers.yml
@@ -30,6 +30,10 @@ en:
         client_has_partner: Does your client have a partner?
       steps_client_details_form:
         date_of_birth: Date of birth
+      steps_client_case_type_form:
+        case_type: What is the case type?
+      steps_client_appeal_details_form:
+        appeal_lodged_date: Date the appeal was lodged
       steps_client_benefit_type_form:
         benefit_type: Does your client get one of these passporting benefits?
       steps_client_contact_details_form:
@@ -38,10 +42,6 @@ en:
         confirm_result: Are the DWP records correct?
       steps_dwp_confirm_details_form:
         confirm_details: Are these details correct?
-      steps_case_case_type_form:
-        case_type: What is the case type?
-      steps_case_appeal_details_form:
-        appeal_lodged_date: Date the appeal was lodged
       steps_case_charges_form:
         date_from: Start date %{index}
         date_to: End date %{index} (optional)
@@ -131,6 +131,18 @@ en:
         first_name: First name
         last_name: Last name
         other_names: Other names (optional)
+      steps_client_case_type_form:
+        case_type_options:
+          summary_only: Summary only
+          either_way: Either way
+          indictable: Indictable
+          already_in_crown_court: Trial already in crown court
+          committal: Committal for sentence
+          appeal_to_crown_court: Appeal to crown court
+          appeal_to_crown_court_with_changes: Appeal to crown court with changes in financial circumstances
+      steps_client_appeal_details_form:
+        appeal_maat_id: Previous MAAT ID (optional)
+        appeal_with_changes_details: What has changed in your client’s financial circumstances?v
       steps_client_has_nino_form:
         nino: Enter your client’s National Insurance number
       steps_client_benefit_type_form:
@@ -169,18 +181,6 @@ en:
         postcode: Postcode
       steps_case_urn_form:
         urn: Enter the unique reference number (URN) for your client’s case (optional)
-      steps_case_case_type_form:
-        case_type_options:
-          summary_only: Summary only
-          either_way: Either way
-          indictable: Indictable
-          already_in_crown_court: Trial already in crown court
-          committal: Committal for sentence
-          appeal_to_crown_court: Appeal to crown court
-          appeal_to_crown_court_with_changes: Appeal to crown court with changes in financial circumstances
-      steps_case_appeal_details_form:
-        appeal_maat_id: Previous MAAT ID (optional)
-        appeal_with_changes_details: What has changed in your client’s financial circumstances?
       steps_case_charges_form:
         offence_name: Offence
       steps_case_charges_summary_form:
