--- spec/controllers/steps/case/appeal_details_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Steps::Case::AppealDetailsController, type: :controller do
-  it_behaves_like 'a generic step controller', Steps::Case::AppealDetailsForm, Decisions::CaseDecisionTree
-end
