+++ spec/controllers/steps/client/appeal_details_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::Client::AppealDetailsController, type: :controller do
+  it_behaves_like 'a generic step controller', Steps::Client::AppealDetailsForm, Decisions::ClientDecisionTree
+end
