+++ spec/controllers/steps/client/date_stamp_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::Client::DateStampController, type: :controller do
+  it_behaves_like 'a no-op advance step controller', :date_stamp, Decisions::ClientDecisionTree
+end
