+++ spec/services/decisions/case_decision_tree_spec.rb
@@ -20,73 +20,6 @@ RSpec.describe Decisions::CaseDecisionTree do
 
   it_behaves_like 'a decision tree'
 
-  context 'when the step is `case_type`' do
-    let(:form_object) { double('FormObject', case: kase, case_type: CaseType.new(case_type)) }
-    let(:step_name) { :case_type }
-
-    context 'and the case type is `appeal_to_crown_court`' do
-      let(:case_type) { CaseType::APPEAL_TO_CROWN_COURT.to_s }
-
-      it { is_expected.to have_destination(:appeal_details, :edit, id: crime_application) }
-    end
-
-    context 'and the case type is `appeal_to_crown_court_with_changes`' do
-      let(:case_type) { CaseType::APPEAL_TO_CROWN_COURT_WITH_CHANGES.to_s }
-
-      it { is_expected.to have_destination(:appeal_details, :edit, id: crime_application) }
-    end
-
-    context 'and the application already has a date stamp' do
-      before do
-        allow(crime_application).to receive(:date_stamp) { Time.zone.today }
-        allow(kase).to receive(:case_type).and_return(case_type)
-      end
-
-      let(:case_type) { CaseType::SUMMARY_ONLY.to_s }
-
-      it { is_expected.to have_destination(:urn, :edit, id: crime_application) }
-    end
-
-    context 'and the application has no date stamp' do
-      before do
-        allow(crime_application).to receive(:date_stamp)
-        allow(kase).to receive(:case_type).and_return(case_type)
-      end
-
-      context 'and the case type is "date stampable"' do
-        let(:charges_double) { double(any?: false, create!: 'charge') }
-        let(:case_type) { CaseType::SUMMARY_ONLY.to_s }
-
-        it { is_expected.to have_destination(:date_stamp, :edit, id: crime_application) }
-      end
-
-      context 'and case type is not "date stampable"' do
-        let(:case_type) { CaseType::INDICTABLE.to_s }
-
-        it { is_expected.to have_destination(:urn, :edit, id: crime_application) }
-      end
-    end
-  end
-
-  context 'when the step is `appeal_details`' do
-    let(:form_object) { double('FormObject') }
-    let(:step_name) { :appeal_details }
-
-    # We've tested this logic for non-appeals, no need to test again
-    # as this step runs the same method/code
-    it 'performs the date stamp logic' do
-      expect(subject).to receive(:date_stamp_if_needed)
-      subject.destination
-    end
-  end
-
-  context 'when the step is `date_stamp`' do
-    let(:form_object) { double('FormObject') }
-    let(:step_name) { :date_stamp }
-
-    it { is_expected.to have_destination(:urn, :edit, id: crime_application) }
-  end
-
   context 'when the step is `urn`' do
     let(:form_object) { double('FormObject') }
     let(:step_name) { :urn }
